tower_http::classify

Enum GrpcCode

#[non_exhaustive]

#[repr(i32)]
pub enum GrpcCode {

    Ok = 0,
    Cancelled = 1,
    Unknown = 2,
    InvalidArgument = 3,
    DeadlineExceeded = 4,
    NotFound = 5,
    AlreadyExists = 6,
    PermissionDenied = 7,
    ResourceExhausted = 8,
    FailedPrecondition = 9,
    Aborted = 10,
    OutOfRange = 11,
    Unimplemented = 12,
    Internal = 13,
    Unavailable = 14,
    DataLoss = 15,
    Unauthenticated = 16,

}

gRPC status codes.

These variants match the gRPC status codes.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Ok = 0

The operation completed successfully.

Cancelled = 1

The operation was cancelled.

Unknown = 2

Unknown error.

InvalidArgument = 3

Client specified an invalid argument.

DeadlineExceeded = 4

Deadline expired before operation could complete.

NotFound = 5

Some requested entity was not found.

AlreadyExists = 6

Some entity that we attempted to create already exists.

PermissionDenied = 7

The caller does not have permission to execute the specified operation.

ResourceExhausted = 8

Some resource has been exhausted.

FailedPrecondition = 9

The system is not in a state required for the operation’s execution.

Aborted = 10

The operation was aborted.

OutOfRange = 11

Operation was attempted past the valid range.

Unimplemented = 12

Operation is not implemented or not supported.

Internal = 13

Internal error.

Unavailable = 14

The service is currently unavailable.

DataLoss = 15

Unrecoverable data loss or corruption.

Unauthenticated = 16

The request does not have valid authentication credentials