Struct tower_http::classify::StatusInRangeAsFailures

pub struct StatusInRangeAsFailures { /* private fields */ }

Response classifier that considers responses with a status code within some range to be failures.

Example

A client with tracing where server errors and client errors are considered failures.

use tower_http::{trace::TraceLayer, classify::StatusInRangeAsFailures};
use tower::{ServiceBuilder, Service, ServiceExt};
use hyper::{Client, Body};
use http::{Request, Method};

let classifier = StatusInRangeAsFailures::new(400..=599);

let mut client = ServiceBuilder::new()
    .layer(TraceLayer::new(classifier.into_make_classifier()))
    .service(Client::new());

let request = Request::builder()
    .method(Method::GET)
    .uri("https://example.com")
    .body(Body::empty())
    .unwrap();

let response = client.ready().await?.call(request).await?;

Implementations

impl StatusInRangeAsFailures

pub fn new(range: RangeInclusive<u16>) -> Self

Creates a new StatusInRangeAsFailures.

Panics

Panics if the start or end of range aren’t valid status codes as determined by StatusCode::from_u16.

pub fn new_for_client_and_server_errors() -> Self

Creates a new StatusInRangeAsFailures that classifies client and server responses as failures.

This is a convenience for StatusInRangeAsFailures::new(400..=599).

pub fn into_make_classifier(self) -> SharedClassifier<Self>

Convert this StatusInRangeAsFailures into a MakeClassifier.