tower_http::compression

Trait Predicate

pub trait Predicate: Clone {
    // Required method
    fn should_compress<B>(&self, response: &Response<B>) -> bool
       where B: Body;

    // Provided method
    fn and<Other>(self, other: Other) -> And<Self, Other>
       where Self: Sized,
             Other: Predicate { ... }
}

Predicate used to determine if a response should be compressed or not.

Required Methods

fn should_compress<B>(&self, response: &Response<B>) -> bool
where B: Body,

Should this response be compressed or not?

Provided Methods

fn and<Other>(self, other: Other) -> And<Self, Other>
where Self: Sized, Other: Predicate,

Combine two predicates into one.

The resulting predicate enables compression if both inner predicates do.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.