Module tower_http::sensitive_header

This is supported on crate feature sensitive-header only.

Middlewares that mark headers as sensitive.

Example

use tower_http::sensitive_header::SetSensitiveHeaderLayer;
use tower::{Service, ServiceExt, ServiceBuilder, service_fn};
use http::{Request, Response, header::AUTHORIZATION};
use hyper::Body;
use std::convert::Infallible;

async fn handle(req: Request<Body>) -> Result<Response<Body>, Infallible> {
    // ...
}

let mut service = ServiceBuilder::new()
    // Mark the `Authorization` header as sensitive so it doesn't show in logs
    //
    // `SetSensitiveHeaderLayer` will mark the header as sensitive on both the
    // request and response.
    .layer(SetSensitiveHeaderLayer::new(AUTHORIZATION))
    .service_fn(handle);

// Call the service.
let response = service
    .ready()
    .await?
    .call(Request::new(Body::empty()))
    .await?;

Structs

SetSensitiveHeaderLayer
Mark a header as sensitive on both requests and responses.

SetSensitiveRequestHeader
Mark a request header as sensitive.

SetSensitiveRequestHeaderLayer
Mark a request header as sensitive.

SetSensitiveResponseHeader
Mark a response header as sensitive.

SetSensitiveResponseHeaderLayer
Mark a response header as sensitive.

SetSensitiveResponseHeaderResponseFuture
Response future for SetSensitiveResponseHeader.

Type Definitions

SetSensitiveHeader
Mark a header as sensitive on both requests and responses.