Module tower_http::services

Services that return responses without wrapping other Services.

These kinds of services are also referred to as “leaf services” since they sit at the leaves of a tree of services.

Modules

fs	fs
File system related services.

redirect	redirect
Service that redirects all requests.

Structs

Redirect
Service that redirects all requests.

ServeDir
Service that serves files from a given directory and all its sub directories.

ServeFile
Service that serves a file.