Struct tower_http::services::fs::ServeDir

pub struct ServeDir<F = DefaultServeDirFallback> { /* private fields */ }

Service that serves files from a given directory and all its sub directories.

The Content-Type will be guessed from the file extension.

An empty response with status 404 Not Found will be returned if:

The file doesn’t exist

Any segment of the path contains ..

Any segment of the path contains a backslash

We don’t have necessary permissions to read the file

Example

use tower_http::services::ServeDir;

// This will serve files in the "assets" directory and
// its subdirectories
let service = ServeDir::new("assets");

// Run our service using `hyper`
let addr = std::net::SocketAddr::from(([127, 0, 0, 1], 3000));
hyper::Server::bind(&addr)
    .serve(tower::make::Shared::new(service))
    .await
    .expect("server error");

Implementations

impl ServeDir<DefaultServeDirFallback>

pub fn new<P>(path: P) -> Self
where P: AsRef<Path>,

Create a new ServeDir.

impl<F> ServeDir<F>

pub fn append_index_html_on_directories(self, append: bool) -> Self

If the requested path is a directory append index.html.

This is useful for static sites.

Defaults to true.

pub fn with_buf_chunk_size(self, chunk_size: usize) -> Self

Set a specific read buffer chunk size.

The default capacity is 64kb.

pub fn precompressed_gzip(self) -> Self

Informs the service that it should also look for a precompressed gzip version of any file in the directory.

Assuming the dir directory is being served and dir/foo.txt is requested, a client with an Accept-Encoding header that allows the gzip encoding will receive the file dir/foo.txt.gz instead of dir/foo.txt. If the precompressed file is not available, or the client doesn’t support it, the uncompressed version will be served instead. Both the precompressed version and the uncompressed version are expected to be present in the directory. Different precompressed variants can be combined.

pub fn precompressed_br(self) -> Self

Informs the service that it should also look for a precompressed brotli version of any file in the directory.

Assuming the dir directory is being served and dir/foo.txt is requested, a client with an Accept-Encoding header that allows the brotli encoding will receive the file dir/foo.txt.br instead of dir/foo.txt. If the precompressed file is not available, or the client doesn’t support it, the uncompressed version will be served instead. Both the precompressed version and the uncompressed version are expected to be present in the directory. Different precompressed variants can be combined.

pub fn precompressed_deflate(self) -> Self

Informs the service that it should also look for a precompressed deflate version of any file in the directory.

Assuming the dir directory is being served and dir/foo.txt is requested, a client with an Accept-Encoding header that allows the deflate encoding will receive the file dir/foo.txt.zz instead of dir/foo.txt. If the precompressed file is not available, or the client doesn’t support it, the uncompressed version will be served instead. Both the precompressed version and the uncompressed version are expected to be present in the directory. Different precompressed variants can be combined.

pub fn fallback<F2>(self, new_fallback: F2) -> ServeDir<F2>

Set the fallback service.

This service will be called if there is no file at the path of the request.

The status code returned by the fallback will not be altered. Use ServeDir::not_found_service to set a fallback and always respond with 404 Not Found.

Example

This can be used to respond with a different file:

use tower_http::services::{ServeDir, ServeFile};

let service = ServeDir::new("assets")
    // respond with `not_found.html` for missing files
    .fallback(ServeFile::new("assets/not_found.html"));

// Run our service using `hyper`
let addr = std::net::SocketAddr::from(([127, 0, 0, 1], 3000));
hyper::Server::bind(&addr)
    .serve(tower::make::Shared::new(service))
    .await
    .expect("server error");

pub fn not_found_service<F2>(self, new_fallback: F2) -> ServeDir<SetStatus<F2>>

Set the fallback service and override the fallback’s status code to 404 Not Found.

This service will be called if there is no file at the path of the request.

Example

This can be used to respond with a different file:

use tower_http::services::{ServeDir, ServeFile};

let service = ServeDir::new("assets")
    // respond with `404 Not Found` and the contents of `not_found.html` for missing files
    .not_found_service(ServeFile::new("assets/not_found.html"));

// Run our service using `hyper`
let addr = std::net::SocketAddr::from(([127, 0, 0, 1], 3000));
hyper::Server::bind(&addr)
    .serve(tower::make::Shared::new(service))
    .await
    .expect("server error");

Setups like this are often found in single page applications.

pub fn call_fallback_on_method_not_allowed(self, call_fallback: bool) -> Self

Customize whether or not to call the fallback for requests that aren’t GET or HEAD.

Defaults to not calling the fallback and instead returning 405 Method Not Allowed.