Struct tower_http::trace::TraceLayer

pub struct TraceLayer<M, MakeSpan = DefaultMakeSpan, OnRequest = DefaultOnRequest, OnResponse = DefaultOnResponse, OnBodyChunk = DefaultOnBodyChunk, OnEos = DefaultOnEos, OnFailure = DefaultOnFailure> { /* private fields */ }

Layer that adds high level tracing to a Service.

See the module docs for more details.

Implementations

impl<M> TraceLayer<M>

pub fn new(make_classifier: M) -> Selfwhere
M: MakeClassifier,

Create a new TraceLayer using the given MakeClassifier.

impl<M, MakeSpan, OnRequest, OnResponse, OnBodyChunk, OnEos, OnFailure> TraceLayer<M, MakeSpan, OnRequest, OnResponse, OnBodyChunk, OnEos, OnFailure>

pub fn on_request<NewOnRequest>(
self,
new_on_request: NewOnRequest
) -> TraceLayer<M, MakeSpan, NewOnRequest, OnResponse, OnBodyChunk, OnEos, OnFailure>

Customize what to do when a request is received.

NewOnRequest is expected to implement OnRequest.

pub fn on_response<NewOnResponse>(
self,
new_on_response: NewOnResponse
) -> TraceLayer<M, MakeSpan, OnRequest, NewOnResponse, OnBodyChunk, OnEos, OnFailure>

Customize what to do when a response has been produced.

NewOnResponse is expected to implement OnResponse.

pub fn on_body_chunk<NewOnBodyChunk>(
self,
new_on_body_chunk: NewOnBodyChunk
) -> TraceLayer<M, MakeSpan, OnRequest, OnResponse, NewOnBodyChunk, OnEos, OnFailure>

Customize what to do when a body chunk has been sent.

NewOnBodyChunk is expected to implement OnBodyChunk.

pub fn on_eos<NewOnEos>(
self,
new_on_eos: NewOnEos
) -> TraceLayer<M, MakeSpan, OnRequest, OnResponse, OnBodyChunk, NewOnEos, OnFailure>

Customize what to do when a streaming response has closed.

NewOnEos is expected to implement OnEos.

pub fn on_failure<NewOnFailure>(
self,
new_on_failure: NewOnFailure
) -> TraceLayer<M, MakeSpan, OnRequest, OnResponse, OnBodyChunk, OnEos, NewOnFailure>

Customize what to do when a response has been classified as a failure.

NewOnFailure is expected to implement OnFailure.

pub fn make_span_with<NewMakeSpan>(
self,
new_make_span: NewMakeSpan
) -> TraceLayer<M, NewMakeSpan, OnRequest, OnResponse, OnBodyChunk, OnEos, OnFailure>

Customize how to make Spans that all request handling will be wrapped in.

NewMakeSpan is expected to implement MakeSpan.

impl TraceLayer<SharedClassifier<ServerErrorsAsFailures>>

pub fn new_for_http() -> Self

Create a new TraceLayer using ServerErrorsAsFailures which supports classifying regular HTTP responses based on the status code.

impl TraceLayer<SharedClassifier<GrpcErrorsAsFailures>>

pub fn new_for_grpc() -> Self

Create a new TraceLayer using GrpcErrorsAsFailures which supports classifying gRPC responses and streams based on the grpc-status header.