Trait tower_http::trace::OnRequest

pub trait OnRequest<B> {
    fn on_request(&mut self, request: &Request<B>, span: &Span);
}

This is supported on crate feature trace only.

Trait used to tell Trace what to do when a request is received.

See the module docs for details on exactly when the on_request callback is called.

Required methods

fn on_request(&mut self, request: &Request<B>, span: &Span)

Do the thing.

span is the tracing Span, corresponding to this request, produced by the closure passed to TraceLayer::make_span_with. It can be used to record field values that weren’t known when the span was created.