Trait tower_http::trace::OnResponse

pub trait OnResponse<B> {
    fn on_response(self, response: &Response<B>, latency: Duration);
}

This is supported on crate feature trace only.

Trait used to tell Trace what to do when a response has been produced.

Required methods

fn on_response(self, response: &Response<B>, latency: Duration)

Do the thing.

latency is the duration since the request was received.

Loading content...