Trait tower_http::validate_request::ValidateRequest

pub trait ValidateRequest<B> {
    type ResponseBody;

    // Required method
    fn validate(
        &mut self,
        request: &mut Request<B>
    ) -> Result<(), Response<Self::ResponseBody>>;
}

Trait for validating requests.

Required Associated Types

type ResponseBody

The body type used for responses to unvalidated requests.

Required Methods

fn validate( &mut self, request: &mut Request<B> ) -> Result<(), Response<Self::ResponseBody>>

Validate the request.

If Ok(()) is returned then the request is allowed through, otherwise not.