Crate tower_http

async fn(HttpRequest) -> Result<HttpResponse, Error>

Overview

tower-http is a library that provides HTTP-specific middlewares and utilities built on top of tower.

All middlewares uses the http and http-body crates as the HTTP abstractions. That means they’re compatible with any library or framework that also uses those crates, such as hyper.

Example server

This example shows how to apply middlewares from tower-http to a Service and then run that service using hyper.

use tower_http::{
    add_extension::AddExtensionLayer,
    compression::CompressionLayer,
    propagate_header::PropagateHeaderLayer,
    sensitive_header::SetSensitiveHeaderLayer,
    set_header::SetResponseHeaderLayer,
};
use tower::{ServiceBuilder, service_fn, make::Shared};
use http::{Request, Response, header::{HeaderName, CONTENT_TYPE, AUTHORIZATION}};
use hyper::{Body, Error, server::Server, service::make_service_fn};
use std::{sync::Arc, net::SocketAddr, convert::Infallible};

// Our request handler. This is where we would implement the application logic
// for responding to HTTP requests...
async fn handler(request: Request<Body>) -> Result<Response<Body>, Error> {
    // ...
}

// Shared state across all request handlers --- in this case, a pool of database connections.
struct State {
    pool: DatabaseConnectionPool,
}

#[tokio::main]
async fn main() {
    // Construct the shared state.
    let state = State {
        pool: DatabaseConnectionPool::new(),
    };

    // Use `tower`'s `ServiceBuilder` API to build a stack of `tower` middleware
    // wrapping our request handler.
    let service = ServiceBuilder::new()
        // Share an `Arc<State>` with all requests
        .layer(AddExtensionLayer::new(Arc::new(state)))
        // Compress responses
        .layer(CompressionLayer::new())
        // Propagate `X-Request-Id`s from requests to responses
        .layer(PropagateHeaderLayer::new(HeaderName::from_static("x-request-id")))
        // Mark the `Authorization` header as sensitive so it doesn't show in logs
        .layer(SetSensitiveHeaderLayer::new(AUTHORIZATION))
        // If the response has a known size set the `Content-Length` header
        .layer(SetResponseHeaderLayer::overriding(CONTENT_TYPE, content_length_from_response))
        // Wrap a `Service` in our middleware stack
        .service_fn(handler);

    // And run our service using `hyper`
    let addr = SocketAddr::from(([127, 0, 0, 1], 3000));
    Server::bind(&addr)
        .serve(Shared::new(service))
        .await
        .expect("server error");
}

Keep in mind that while this example uses hyper, tower-http supports any HTTP client/server implementation that uses the http and http-body crates.

Example client

tower-http middlewares can also be applied to HTTP clients:

use tower_http::{
    decompression::DecompressionLayer,
    set_header::SetRequestHeaderLayer,
};
use tower::{ServiceBuilder, Service, ServiceExt};
use hyper::Body;
use http::{Request, Response, HeaderValue, header::USER_AGENT};

#[tokio::main]
async fn main() {
    let mut client = ServiceBuilder::new()
        // Set a `User-Agent` header on all requests
        .layer(SetRequestHeaderLayer::<_, Body>::overriding(
            USER_AGENT,
            HeaderValue::from_static("tower-http demo")
        ))
        // Decompress response bodies
        .layer(DecompressionLayer::new())
        // Wrap a `hyper::Client` in our middleware stack
        .service(hyper::Client::new());

    // Make a request
    let request = Request::builder()
        .uri("http://example.com")
        .body(Body::empty())
        .unwrap();

    let response = client
        .ready()
        .await
        .unwrap()
        .call(request)
        .await
        .unwrap();
}

Feature Flags

All middleware are disabled by default and can be enabled using cargo features.

For example, to enable the AddExtension middleware, add the “add-extension” feature flag in yourCargo.toml:

tower-http = { version = "0.1.0", features = ["add-extension"] }

You can use "full" to enable everything:

tower-http = { version = "0.1.0", features = ["full"] }

Modules

add_extension	add-extension
Middleware that clones a value into each request’s extensions.

classify
Tools for classifying responses as either success or failure.

compression	compression
Middleware that compresses response bodies.

decompression	decompression
Middleware that decompresses response bodies.

follow_redirect	follow-redirect
Middleware for following redirections.

map_request_body	map-request-body
Apply a transformation to the request body.

map_response_body	map-response-body
Apply a transformation to the response body.

propagate_header	propagate-header
Propagate a header from the request to the response.

sensitive_header	sensitive-header
Middlewares that mark headers as sensitive.

services
Services that return responses without wrapping other Services.

set_header	set-header
Middlewares for setting headers on requests and responses.

trace	trace
Middleware that adds high level tracing to a Service.

Enums

BodyOrIoError	compression or decompression
Error type containing either a body error or an IO error.

LatencyUnit
The latency unit used to report latencies by middlewares.