tower_http

Module auth

Authorization related middleware.

Modules

add_authorization

Add authorization to requests using the Authorization header.

async_require_authorization

Authorize requests using the Authorization header asynchronously.

require_authorizationDeprecated

Authorize requests using ValidateRequest.

Structs

AddAuthorization

Middleware that adds authorization all requests using the Authorization header.

AddAuthorizationLayer

Layer that applies AddAuthorization which adds authorization to all requests using the Authorization header.

AsyncRequireAuthorization

Middleware that authorizes all requests using the Authorization header.

AsyncRequireAuthorizationLayer

Layer that applies AsyncRequireAuthorization which authorizes all requests using the Authorization header.

Traits

AsyncAuthorizeRequest

Trait for authorizing requests.