Module tower_http::classify

Tools for classifying responses as either success or failure.

Structs

GrpcEosErrorsAsFailures
The ClassifyEos for GrpcErrorsAsFailures.

GrpcErrorsAsFailures
Response classifier for gRPC responses.

NeverClassifyEos
A ClassifyEos type that can be used in ClassifyResponse implementations that never have to classify streaming responses.

ServerErrorsAsFailures
The default classifier used for normal HTTP responses.

SharedClassifier
A MakeClassifier that produces new classifiers by cloning an inner classifier.

Enums

ClassifiedResponse
Result of doing a classification.

Traits

ClassifyEos
Trait for classifying end of streams (EOS) as either success or failure.

ClassifyResponse
Trait for classifying responses as either success or failure. Designed to support both unary requests (single request for a single response) as well as streaming responses.

MakeClassifier
Trait for producing response classifiers from a request.