Enum tower_http::classify::GrpcCode

pub enum GrpcCode {

    Ok,
    Cancelled,
    Unknown,
    InvalidArgument,
    DeadlineExceeded,
    NotFound,
    AlreadyExists,
    PermissionDenied,
    ResourceExhausted,
    FailedPrecondition,
    Aborted,
    OutOfRange,
    Unimplemented,
    Internal,
    Unavailable,
    DataLoss,
    Unauthenticated,

}

gRPC status codes. Used in GrpcErrorsAsFailures::success_codes.

These variants match the gRPC status codes.

Variants

Ok

The operation completed successfully.

Cancelled

The operation was cancelled.

Unknown

Unknown error.

InvalidArgument

Client specified an invalid argument.

DeadlineExceeded

Deadline expired before operation could complete.

NotFound

Some requested entity was not found.

AlreadyExists

Some entity that we attempted to create already exists.

PermissionDenied

The caller does not have permission to execute the specified operation.

ResourceExhausted

Some resource has been exhausted.

FailedPrecondition

The system is not in a state required for the operation’s execution.

Aborted

The operation was aborted.

OutOfRange

Operation was attempted past the valid range.

Unimplemented

Operation is not implemented or not supported.

Internal

Internal error.

Unavailable

The service is currently unavailable.

DataLoss

Unrecoverable data loss or corruption.

Unauthenticated

The request does not have valid authentication credentials