Trait tower_http::classify::ClassifyEos

pub trait ClassifyEos<E> {
    type FailureClass;
    fn classify_eos(
        self, 
        trailers: Option<&HeaderMap>
    ) -> Result<(), Self::FailureClass>;

    fn classify_error(self, error: &E) -> Self::FailureClass;
}

Trait for classifying end of streams (EOS) as either success or failure.

Associated Types

type FailureClass

The type of failure classifications.

Loading content...
Required methods

fn classify_eos(
self,
trailers: Option<&HeaderMap>
) -> Result<(), Self::FailureClass>

Perform the classification from response trailers.

fn classify_error(self, error: &E) -> Self::FailureClass

Classify an error.

Errors are always errors (doh) but sometimes it might be useful to have multiple classes of errors. A retry policy might allow retrying some errors and not others.

Loading content...