Module tower_http::compression::predicate

Predicates for disabling compression of responses.

Predicates are applied with Compression::compress_when or CompressionLayer::compress_when.

Structs

And

Two predicates combined into one.

DefaultPredicate

The default predicate used by Compression and CompressionLayer.

NotForContentType

Predicate that wont allow responses with a specific content-type to be compressed.

SizeAbove

Predicate that will only allow compression of responses above a certain size.

Traits

Predicate

Predicate used to determine if a response should be compressed or not.