tower_http

Module cors

Middleware which adds headers for CORS.

Example

use http::{Request, Response, Method, header};
use http_body_util::Full;
use bytes::Bytes;
use tower::{ServiceBuilder, ServiceExt, Service};
use tower_http::cors::{Any, CorsLayer};
use std::convert::Infallible;

async fn handle(request: Request<Full<Bytes>>) -> Result<Response<Full<Bytes>>, Infallible> {
    Ok(Response::new(Full::default()))
}

let cors = CorsLayer::new()
    // allow `GET` and `POST` when accessing the resource
    .allow_methods([Method::GET, Method::POST])
    // allow requests from any origin
    .allow_origin(Any);

let mut service = ServiceBuilder::new()
    .layer(cors)
    .service_fn(handle);

let request = Request::builder()
    .header(header::ORIGIN, "https://example.com")
    .body(Full::default())
    .unwrap();

let response = service
    .ready()
    .await?
    .call(request)
    .await?;

assert_eq!(
    response.headers().get(header::ACCESS_CONTROL_ALLOW_ORIGIN).unwrap(),
    "*",
);

Structs

AllowCredentials

Holds configuration for how to set the Access-Control-Allow-Credentials header.

AllowHeaders

Holds configuration for how to set the Access-Control-Allow-Headers header.

AllowMethods

Holds configuration for how to set the Access-Control-Allow-Methods header.

AllowOrigin

Holds configuration for how to set the Access-Control-Allow-Origin header.

AllowPrivateNetwork

Holds configuration for how to set the Access-Control-Allow-Private-Network header.

Any

Represents a wildcard value (*) used with some CORS headers such as CorsLayer::allow_methods.

Cors

Middleware which adds headers for CORS.

CorsLayer

Layer that applies the Cors middleware which adds headers for CORS.

ExposeHeaders

Holds configuration for how to set the Access-Control-Expose-Headers header.

MaxAge

Holds configuration for how to set the Access-Control-Max-Age header.

ResponseFuture

Response future for Cors.

Vary

Holds configuration for how to set the Vary header.

Functions

anyDeprecated

Represents a wildcard value (*) used with some CORS headers such as CorsLayer::allow_methods.

preflight_request_headers

Returns an iterator over the three request headers that may be involved in a CORS preflight request.