tower_http

Module services

Services that return responses without wrapping other Services.

These kinds of services are also referred to as “leaf services” since they sit at the leaves of a tree of services.