Struct tower_http::trace::Trace

pub struct Trace<S, M, MakeSpan = DefaultMakeSpan, OnRequest = DefaultOnRequest, OnResponse = DefaultOnResponse, OnBodyChunk = DefaultOnBodyChunk, OnEos = DefaultOnEos, OnFailure = DefaultOnFailure> { /* fields omitted */ }

Middleware that adds high level tracing to a Service.

See the module docs for an example.

Implementations

impl<S, M> Trace<S, M>

pub fn new(inner: S, make_classifier: M) -> Self where
M: MakeClassifier,

Create a new Trace using the given MakeClassifier.

pub fn layer(make_classifier: M) -> TraceLayer<M> where
M: MakeClassifier,

Returns a new Layer that wraps services with a TraceLayer middleware.

impl<S, M, MakeSpan, OnRequest, OnResponse, OnBodyChunk, OnEos, OnFailure> Trace<S, M, MakeSpan, OnRequest, OnResponse, OnBodyChunk, OnEos, OnFailure>

pub fn get_ref(&self) -> &S

Gets a reference to the underlying service.

pub fn get_mut(&mut self) -> &mut S

Gets a mutable reference to the underlying service.

pub fn into_inner(self) -> S

Consumes self, returning the underlying service.

pub fn on_request<NewOnRequest>(
self,
new_on_request: NewOnRequest
) -> Trace<S, M, MakeSpan, NewOnRequest, OnResponse, OnBodyChunk, OnEos, OnFailure>

Customize what to do when a request is received.

NewOnRequest is expected to implement OnRequest.

pub fn on_response<NewOnResponse>(
self,
new_on_response: NewOnResponse
) -> Trace<S, M, MakeSpan, OnRequest, NewOnResponse, OnBodyChunk, OnEos, OnFailure>

Customize what to do when a response has been produced.

NewOnResponse is expected to implement OnResponse.

pub fn on_body_chunk<NewOnBodyChunk>(
self,
new_on_body_chunk: NewOnBodyChunk
) -> Trace<S, M, MakeSpan, OnRequest, OnResponse, NewOnBodyChunk, OnEos, OnFailure>

Customize what to do when a body chunk has been sent.

NewOnBodyChunk is expected to implement OnBodyChunk.

pub fn on_eos<NewOnEos>(
self,
new_on_eos: NewOnEos
) -> Trace<S, M, MakeSpan, OnRequest, OnResponse, OnBodyChunk, NewOnEos, OnFailure>

Customize what to do when a streaming response has closed.

NewOnEos is expected to implement OnEos.

pub fn on_failure<NewOnFailure>(
self,
new_on_failure: NewOnFailure
) -> Trace<S, M, MakeSpan, OnRequest, OnResponse, OnBodyChunk, OnEos, NewOnFailure>

Customize what to do when a response has been classified as a failure.

NewOnFailure is expected to implement OnFailure.

pub fn make_span_with<NewMakeSpan>(
self,
new_make_span: NewMakeSpan
) -> Trace<S, M, NewMakeSpan, OnRequest, OnResponse, OnBodyChunk, OnEos, OnFailure>

Customize how to make Spans that all request handling will be wrapped in.

NewMakeSpan is expected to implement MakeSpan.

impl<S> Trace<S, SharedClassifier<ServerErrorsAsFailures>, DefaultMakeSpan, DefaultOnRequest, DefaultOnResponse, DefaultOnBodyChunk, DefaultOnEos, DefaultOnFailure>

pub fn new_for_http(inner: S) -> Self

Create a new Trace using ServerErrorsAsFailures which supports classifying regular HTTP responses based on the status code.

impl<S> Trace<S, SharedClassifier<GrpcErrorsAsFailures>, DefaultMakeSpan, DefaultOnRequest, DefaultOnResponse, DefaultOnBodyChunk, DefaultOnEos, DefaultOnFailure>

pub fn new_for_grpc(inner: S) -> Self

Create a new Trace using GrpcErrorsAsFailures which supports classifying gRPC responses and streams based on the grpc-status header.