tower_http::trace

Trait OnEos

pub trait OnEos {
    // Required method
    fn on_eos(
        self,
        trailers: Option<&HeaderMap>,
        stream_duration: Duration,
        span: &Span,
    );
}

Trait used to tell Trace what to do when a stream closes.

See the module docs for details on exactly when the on_eos callback is called.

Required Methods

fn on_eos( self, trailers: Option<&HeaderMap>, stream_duration: Duration, span: &Span, )

Do the thing.

stream_duration is the duration since the response was sent.

span is the tracing Span, corresponding to this request, produced by the closure passed to TraceLayer::make_span_with. It can be used to record field values that weren’t known when the span was created.