Trait tower_http::trace::OnResponse

pub trait OnResponse<B> {
    // Required method
    fn on_response(self, response: &Response<B>, latency: Duration, span: &Span);
}

Trait used to tell Trace what to do when a response has been produced.

See the module docs for details on exactly when the on_response callback is called.

Required Methods

fn on_response(self, response: &Response<B>, latency: Duration, span: &Span)

Do the thing.

latency is the duration since the request was received.

span is the tracing Span, corresponding to this request, produced by the closure passed to TraceLayer::make_span_with. It can be used to record field values that weren’t known when the span was created.