Trait tower_http::ServiceBuilderExt

pub trait ServiceBuilderExt<L>: Sealed<L> + Sized {
    // Required methods
    fn map_request_body<F>(
        self,
        f: F
    ) -> ServiceBuilder<Stack<MapRequestBodyLayer<F>, L>>;
    fn follow_redirects(
        self
    ) -> ServiceBuilder<Stack<FollowRedirectLayer<Standard>, L>>;
    fn override_request_header<M>(
        self,
        header_name: HeaderName,
        make: M
    ) -> ServiceBuilder<Stack<SetRequestHeaderLayer<M>, L>>;
    fn append_request_header<M>(
        self,
        header_name: HeaderName,
        make: M
    ) -> ServiceBuilder<Stack<SetRequestHeaderLayer<M>, L>>;
    fn insert_request_header_if_not_present<M>(
        self,
        header_name: HeaderName,
        make: M
    ) -> ServiceBuilder<Stack<SetRequestHeaderLayer<M>, L>>;
    fn override_response_header<M>(
        self,
        header_name: HeaderName,
        make: M
    ) -> ServiceBuilder<Stack<SetResponseHeaderLayer<M>, L>>;
    fn append_response_header<M>(
        self,
        header_name: HeaderName,
        make: M
    ) -> ServiceBuilder<Stack<SetResponseHeaderLayer<M>, L>>;
    fn insert_response_header_if_not_present<M>(
        self,
        header_name: HeaderName,
        make: M
    ) -> ServiceBuilder<Stack<SetResponseHeaderLayer<M>, L>>;
}

Extension trait that adds methods to tower::ServiceBuilder for adding middleware from tower-http.

Example

use http::{Request, Response, header::HeaderName};
use hyper::Body;
use std::{time::Duration, convert::Infallible};
use tower::{ServiceBuilder, ServiceExt, Service};
use tower_http::ServiceBuilderExt;

async fn handle(request: Request<Body>) -> Result<Response<Body>, Infallible> {
    Ok(Response::new(Body::empty()))
}

let service = ServiceBuilder::new()
    // Methods from tower
    .timeout(Duration::from_secs(30))
    // Methods from tower-http
    .trace_for_http()
    .compression()
    .propagate_header(HeaderName::from_static("x-request-id"))
    .service_fn(handle);

Required Methods

fn map_request_body<F>( self, f: F ) -> ServiceBuilder<Stack<MapRequestBodyLayer<F>, L>>

Apply a transformation to the request body.

See tower_http::map_request_body for more details.

fn follow_redirects( self ) -> ServiceBuilder<Stack<FollowRedirectLayer<Standard>, L>>

Follow redirect resposes using the Standard policy.

See tower_http::follow_redirect for more details.

fn override_request_header<M>( self, header_name: HeaderName, make: M ) -> ServiceBuilder<Stack<SetRequestHeaderLayer<M>, L>>

Insert a header into the request.

If a previous value exists for the same header, it is removed and replaced with the new header value.

See tower_http::set_header for more details.

fn append_request_header<M>( self, header_name: HeaderName, make: M ) -> ServiceBuilder<Stack<SetRequestHeaderLayer<M>, L>>

Append a header into the request.

If previous values exist, the header will have multiple values.

See tower_http::set_header for more details.

fn insert_request_header_if_not_present<M>( self, header_name: HeaderName, make: M ) -> ServiceBuilder<Stack<SetRequestHeaderLayer<M>, L>>

Insert a header into the request, if the header is not already present.

See tower_http::set_header for more details.

fn override_response_header<M>( self, header_name: HeaderName, make: M ) -> ServiceBuilder<Stack<SetResponseHeaderLayer<M>, L>>

Insert a header into the response.

If a previous value exists for the same header, it is removed and replaced with the new header value.

See tower_http::set_header for more details.

fn append_response_header<M>( self, header_name: HeaderName, make: M ) -> ServiceBuilder<Stack<SetResponseHeaderLayer<M>, L>>

Append a header into the response.

If previous values exist, the header will have multiple values.

See tower_http::set_header for more details.

fn insert_response_header_if_not_present<M>( self, header_name: HeaderName, make: M ) -> ServiceBuilder<Stack<SetResponseHeaderLayer<M>, L>>

Insert a header into the response, if the header is not already present.

See tower_http::set_header for more details.