tower_http

Trait ServiceExt

pub trait ServiceExt {
    // Provided methods
    fn trace_for_http(self) -> Trace<Self, HttpMakeClassifier>
       where Self: Sized { ... }
    fn trace_for_grpc(self) -> Trace<Self, GrpcMakeClassifier>
       where Self: Sized { ... }
    fn set_request_id<M>(
        self,
        header_name: HeaderName,
        make_request_id: M,
    ) -> SetRequestId<Self, M>
       where Self: Sized,
             M: MakeRequestId { ... }
    fn set_x_request_id<M>(self, make_request_id: M) -> SetRequestId<Self, M>
       where Self: Sized,
             M: MakeRequestId { ... }
    fn propagate_request_id(
        self,
        header_name: HeaderName,
    ) -> PropagateRequestId<Self>
       where Self: Sized { ... }
    fn propagate_x_request_id(self) -> PropagateRequestId<Self>
       where Self: Sized { ... }
}

Extension trait that adds methods to any Service for adding middleware from tower-http.

Provided Methods

fn trace_for_http(self) -> Trace<Self, HttpMakeClassifier>
where Self: Sized,

High level tracing that classifies responses using HTTP status codes.

This method does not support customizing the output, to do that use TraceLayer instead.

See tower_http::trace for more details.

fn trace_for_grpc(self) -> Trace<Self, GrpcMakeClassifier>
where Self: Sized,

High level tracing that classifies responses using gRPC headers.

This method does not support customizing the output, to do that use TraceLayer instead.

See tower_http::trace for more details.

fn set_request_id<M>( self, header_name: HeaderName, make_request_id: M, ) -> SetRequestId<Self, M>
where Self: Sized, M: MakeRequestId,

Add request id header and extension.

See tower_http::request_id for more details.

fn set_x_request_id<M>(self, make_request_id: M) -> SetRequestId<Self, M>
where Self: Sized, M: MakeRequestId,

Add request id header and extension, using x-request-id as the header name.

See tower_http::request_id for more details.

fn propagate_request_id( self, header_name: HeaderName, ) -> PropagateRequestId<Self>
where Self: Sized,

Propgate request ids from requests to responses.

See tower_http::request_id for more details.

fn propagate_x_request_id(self) -> PropagateRequestId<Self>
where Self: Sized,

Propgate request ids from requests to responses, using x-request-id as the header name.

See tower_http::request_id for more details.